chore: enhance evaluation prompt preview and variable mapping functionality with new hooks and components
